Fixed a bug in the subword parser where reduction with epsilon production were incorrectly returned by getSuccessorActions().
I have also found out that the loop detection improvements commited in
ca7f435ace97cd2511bc59ffde03a8d043a98c3b were more than a mere simplification:
the previous algorithm was ***incorrect*** and did not always detect loops when
they exist.  It also could falsely detect a loop (although that is probably
somewhat uncommon occurrence).  The reason behind the bug is that the previous
algorithm did not respect the boundaries of the current parsing subtree and
took into account reductions from the subtree on the left of the current
subtree.